Simplify development/building docker images
Combine all sawtooth-dev and sawtooth-build Dockerfiles for each language, so
there is only one file to manage. This image is tagged multiple times for
compatibility.

Developers wanting to work in docker should now be able to just run the
sawtooth-dev docker image for the language they are working in and have access
to everything needed by that language's components.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>