Fix ctrs: 1. SGX: Bind counter to signup
Per PoET spec, the monotonic counter must be part of the sealed signup data.
It ensures that only a single enclave can produce wait timers for a validator
registration (embodied by a signup data).

Trying to track the MCID and other items in enclave state does not preclude
certain errors. Instead it is best to pass the signup data into the two calls
which rely on the poet keys and hardware counter for security
(i.e. ecall_CreateWaitTimer and ecall_createWaitCertificate).

Whenever a wait timer is created we increment the counter.

Whenever a wait certificate is requested we check if the counter in the wait
timer matches the value currently in the monotonic counter.

Whenever a wait certificate is created we increment the counter.

Changes:
Add the counter ID to the signup data structure.
Remove reliance on poet state object.
Add counter tracking behavior to wait timer and certificate functions.

Signed-off-by: Dan Middleton <dan.middleton@intel.com>