Fix HW Counter management. Release old signups.
Hardware counters must be released when no longer needed.
This change adds a simple policy to poet to release counters
after a validator registration has exhausted its allowed
number of blocks published (k test).

The SGX enclave is also updated to throw an error if PoET
attempts to unseal signup data that references a defunct counter.

The block publisher now traps for that event to no longer use
that signup data and puts it into the condition where it seeks
to sign up again.

Signed-off-by: Dan Middleton <dan.middleton@intel.com>