Fix misc lint errors (again)
pylint 1.8 was released, and of course that means it now does a few
more checks. It also reveals violations that it didn't find
before (which I guess means it has improved). Just to be safe, I also
ran flake8 over the whole repo, which caught a few more infelicities.

This commit mostly covers whitespace irregularites and string
formatting issues (apparently `%s` should be used in logging
statements, but `.format` should be used for exceptions).

Signed-off-by: Nick Drozd <drozd@bitwise.io>